+++ src/debug-overlay/tabs/InspectorTab.js
@@ -13,7 +13,6 @@ import ConsoleShell from '../../topology-console/components/ConsoleShell';
 import { NewNodeModal } from '../../topology-console/components/Modal';
 import { useJsCatalog } from '../../topology-console/hooks/useJsCatalog';
 import { useClassCatalog } from '../../topology-console/hooks/useClassCatalog';
-import { useVaults } from '../../topology-console/hooks/useVaults';
 import { ShellNode } from '../../runtime/shell-node';
 import { useNodeState } from '../../runtime/react';
 import { useCompletion } from '../../topology-console/hooks/useCompletion';
@@ -158,7 +157,6 @@ export default function InspectorTab( {
 	const jsCatalog = useJsCatalog();
 	const phpCatalog = useClassCatalog( { enabled: !! cwd } );
 	const catalog = cwd ? phpCatalog : jsCatalog;
-	const vaultCatalog = useVaults( { enabled: !! cwd } );
 	const {
 		graph,
 		ready: graphHasNodes,
@@ -310,7 +308,6 @@ export default function InspectorTab( {
 					classCatalog={ schemasByShellName }
 					classes={ catalog.classes }
 					formatters={ catalog.formatters }
-					vaults={ vaultCatalog.vaults }
 					composeTargets={ composeTargets }
 				>
 					<div